src/main/java/folk/tradingbot/Utils.java
package folk.tradingbot;

import java.io.BufferedReader;
import java.io.IOException;
import java.io.InputStreamReader;

public class Utils {
    public static String printPromptAndGetAns(String prompt) {
        System.out.print(prompt);
        BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
        String str = "";
        try {
            str = reader.readLine();
        } catch (IOException e) {
            e.printStackTrace();
        }
        return str;
    }

    public static void sleep(int seconds) {
        if (seconds > 60)
            throw new RuntimeException("Медот Sleep будет ждать больше минуты. Не очень хорошо");
        try {
            Thread.sleep(seconds * 1000L);
        } catch (InterruptedException e) {
            e.printStackTrace();
        }
    }
}
